+++ otx/algorithms/segmentation/adapters/mmseg/configurer.py
@@ -572,10 +572,6 @@ class SemiSLSegmentationConfigurer(SegmentationConfigurer):
         """Adjust settings for task adaptation."""
         super().configure_task(cfg, training, **kwargs)
 
-        # Don't pass task_adapt arg to semi-segmentor
-        if cfg.model.type != "ClassIncrEncoderDecoder" and cfg.model.get("task_adapt", False):
-            cfg.model.pop("task_adapt")
-
         # Remove task adapt hook (set default torch random sampler)
         remove_custom_hook(cfg, "TaskAdaptHook")