+++ otx/algorithms/segmentation/configs/ocr_lite_hrnet_18_mod2/semisl/model.py
@@ -29,7 +29,7 @@ model = dict(
     test_cfg=dict(mode="whole", output_scale=5.0),
     pretrained=None,
     decode_head=dict(
-        type="FCNHead",
+        type="CustomFCNHead",
         in_channels=[40, 80, 160, 320],
         in_index=[0, 1, 2, 3],
         input_transform="multiple_select",
@@ -42,7 +42,6 @@ model = dict(
         norm_cfg=dict(type="BN", requires_grad=True),
         align_corners=False,
         enable_aggregator=True,
-        enable_out_norm=False,
         loss_decode=[
             dict(
                 type="CrossEntropyLoss",